[#15] Move rating item to its own component
